pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-videopipeline"
version = "0.1.0"
description = "AI-driven video production pipeline: text → video → multi-platform publish"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "typer[all]>=0.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]
tts = [
    "edge-tts>=6.1",
]
video = [
    "runwayml>=4.0",
]
all = [
    "edge-tts>=6.1",
    "runwayml>=4.0",
]

[project.scripts]
videopipeline = "videopipeline.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/videopipeline"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
